Conflicts:
	classes/auto-package-utils.bbclass
	classes/base.bbclass
	classes/kernel-headers.bbclass
	classes/kernel.bbclass
	classes/u-boot.bbclass
	conf/distro/core.conf
	conf/provided/all.conf
	conf/u-boot.conf
	lib/oelite/baker.py
	lib/oelite/db.py
	lib/oelite/runq.py
	recipes/crosstool-ng/crosstool-ng-1.9.3/gcc-4.5.2.patch
	recipes/crosstool-ng/crosstool-ng-1.9.3/glibc-2.11.2.patch
	recipes/crosstool-ng/crosstool-ng-1.9.3/glibc-2.9-typedef-caddr.patch
	recipes/crosstool-ng/crosstool-ng/glibc-march-i686.patch
	recipes/crosstool-ng/crosstool-ng_1.9.3.bb
	recipes/crosstool-ng/gcc_1.9.3.bb
	recipes/crosstool-ng/toolchain.inc

+++ classes/tar-image.bbclass
@@ -7,18 +7,24 @@ TAR_IMAGE_DIRNAME ?= "${IMAGE_BASENAME}"
 
 tar_image () {
 	(
-	cd `dirname ${IMAGE_DIR}`
-	imagedir=`basename ${IMAGE_DIR}`
-	rm -rf $imagedir.tmp
-	mkdir $imagedir.tmp
-	mv $imagedir $imagedir.tmp/${TAR_IMAGE_DIRNAME}
-	ln -s $imagedir.tmp/${TAR_IMAGE_DIRNAME} $imagedir
-	tar c -C $imagedir.tmp \
-		-a -f ${B}/${IMAGE_BASENAME}.${TAR_IMAGE_EXT} \
-		${TAR_IMAGE_DIRNAME}
-	rm $imagedir
-	mv $imagedir.tmp/${TAR_IMAGE_DIRNAME} $imagedir
-	rmdir $imagedir.tmp
+	if [ -n "${TAR_IMAGE_DIRNAME}" ]; then
+		cd `dirname ${IMAGE_DIR}`
+		imagedir=`basename ${IMAGE_DIR}`
+		rm -rf $imagedir.tmp
+		mkdir $imagedir.tmp
+		mv $imagedir $imagedir.tmp/${TAR_IMAGE_DIRNAME}
+		ln -s $imagedir.tmp/${TAR_IMAGE_DIRNAME} $imagedir
+		tar c -C $imagedir.tmp \
+			-a -f ${B}/${IMAGE_BASENAME}.${TAR_IMAGE_EXT} \
+			${TAR_IMAGE_DIRNAME}
+		rm $imagedir
+		mv $imagedir.tmp/${TAR_IMAGE_DIRNAME} $imagedir
+		rmdir $imagedir.tmp
+        else
+		tar c -C ${IMAGE_DIR} \
+			-a -f ${B}/${IMAGE_BASENAME}.${TAR_IMAGE_EXT} \
+			.
+	fi
 	)
 }
